+++ tests/test_message.py
@@ -76,7 +76,7 @@ def test_message_auto_body_type():
 
 def test_message_body_data_type():
     single_data = b'!@#$%^&*()_+1234567890'
-    single_data_message = Message(body=single_data, body_type=MessageBodyType.DataType)
+    single_data_message = Message(body=single_data, body_type=MessageBodyType.Data)
     check_list = [data for data in single_data_message.get_data()]
     assert isinstance(single_data_message._body, DataBody)
     assert len(check_list) == 1
@@ -84,7 +84,7 @@ def test_message_body_data_type():
     assert str(single_data_message)
 
     multiple_data = [b'!@#$%^&*()_+1234567890', 'abcdefg~123',]
-    multiple_data_message = Message(body=multiple_data, body_type=MessageBodyType.DataType)
+    multiple_data_message = Message(body=multiple_data, body_type=MessageBodyType.Data)
     check_list = [data for data in multiple_data_message.get_data()]
     assert isinstance(multiple_data_message._body, DataBody)
     assert len(check_list) == 2
@@ -93,39 +93,39 @@ def test_message_body_data_type():
     assert str(multiple_data_message)
 
     with pytest.raises(TypeError):
-        Message(body={"key": "value"}, body_type=MessageBodyType.DataType)
+        Message(body={"key": "value"}, body_type=MessageBodyType.Data)
 
     with pytest.raises(TypeError):
-        Message(body=1, body_type=MessageBodyType.DataType)
+        Message(body=1, body_type=MessageBodyType.Data)
 
     with pytest.raises(TypeError):
-        Message(body=['abc', True], body_type=MessageBodyType.DataType)
+        Message(body=['abc', True], body_type=MessageBodyType.Data)
 
     with pytest.raises(TypeError):
-        Message(body=True, body_type=MessageBodyType.DataType)
+        Message(body=True, body_type=MessageBodyType.Data)
 
 
 def test_message_body_value_type():
     string_value = b'!@#$%^&*()_+1234567890'
-    string_value_message = Message(body=string_value, body_type=MessageBodyType.ValueType)
+    string_value_message = Message(body=string_value, body_type=MessageBodyType.Value)
     assert string_value_message.get_data() == string_value
     assert isinstance(string_value_message._body, ValueBody)
     assert str(string_value_message)
 
     float_value = 1.23
-    float_value_message = Message(body=float_value, body_type=MessageBodyType.ValueType)
+    float_value_message = Message(body=float_value, body_type=MessageBodyType.Value)
     assert float_value_message.get_data() == float_value
     assert isinstance(string_value_message._body, ValueBody)
     assert str(float_value_message)
 
     dict_value = {b"key1 ": b"value1", b'key2': -1, b'key3': False}
-    dict_value_message = Message(body=dict_value, body_type=MessageBodyType.ValueType)
+    dict_value_message = Message(body=dict_value, body_type=MessageBodyType.Value)
     assert dict_value_message.get_data() == dict_value
     assert isinstance(string_value_message._body, ValueBody)
     assert str(dict_value_message)
 
     compound_list_value = [1, b'abc', True, [1.23, b'abc', False], {b"key1 ": b"value1", b'key2': -1}]
-    compound_list_value_message = Message(body=compound_list_value, body_type=MessageBodyType.ValueType)
+    compound_list_value_message = Message(body=compound_list_value, body_type=MessageBodyType.Value)
     assert compound_list_value_message.get_data() == compound_list_value
     assert isinstance(string_value_message._body, ValueBody)
     assert str(compound_list_value_message)
@@ -134,7 +134,7 @@ def test_message_body_value_type():
 def test_message_body_sequence_type():
 
     single_list = [1, 2, b'aa', b'bb', True, b"abc", {b"key1": b"value", b"key2": -1.23}]
-    single_list_message = Message(body=single_list, body_type=MessageBodyType.SequenceType)
+    single_list_message = Message(body=single_list, body_type=MessageBodyType.Sequence)
     check_list = [data for data in single_list_message.get_data()]
     assert isinstance(single_list_message._body, SequenceBody)
     assert len(check_list) == 1
@@ -142,7 +142,7 @@ def test_message_body_sequence_type():
     assert str(single_list_message)
 
     multiple_lists = [[1, 2, 3, 4], [b'aa', b'bb', b'cc', b'dd']]
-    multiple_lists_message = Message(body=multiple_lists, body_type=MessageBodyType.SequenceType)
+    multiple_lists_message = Message(body=multiple_lists, body_type=MessageBodyType.Sequence)
     check_list = [data for data in multiple_lists_message.get_data()]
     assert isinstance(multiple_lists_message._body, SequenceBody)
     assert len(check_list) == 2
@@ -151,13 +151,13 @@ def test_message_body_sequence_type():
     assert str(multiple_lists_message)
 
     with pytest.raises(TypeError):
-        Message(body={"key": "value"}, body_type=MessageBodyType.SequenceType)
+        Message(body={"key": "value"}, body_type=MessageBodyType.Sequence)
 
     with pytest.raises(TypeError):
-        Message(body=1, body_type=MessageBodyType.SequenceType)
+        Message(body=1, body_type=MessageBodyType.Sequence)
 
     with pytest.raises(TypeError):
-        Message(body='a', body_type=MessageBodyType.SequenceType)
+        Message(body='a', body_type=MessageBodyType.Sequence)
 
     with pytest.raises(TypeError):
-        Message(body=True, body_type=MessageBodyType.SequenceType)
+        Message(body=True, body_type=MessageBodyType.Sequence)