+++ HISTORY.rst
@@ -3,6 +3,16 @@
 Release History
 ===============
 
+1.3.0 (Unreleased)
+
+- Added support for AMQP Sequence as the body type of an amqp message.
+- Added new class `uamqp.MessageBodyType` to represent the body type of an amqp message, including:
+  - `Data`: The body consists of one or more data sections and each section contains opaque binary data.
+  - `Sequence`: The body consists of one or more sequence sections and each section contains an arbitrary number of structured data elements.
+  - `Value`: The body consists of one amqp-value section and the section contains a single AMQP value.
+- Added new parameter `body_type` to the constructor of `uamqp.Message` which takes `uamqp.MessageBodyType` to specify the body type of an amqp message.
+- Added new parameter `footer` to the constructor of  `uamqp.Message` which takes a dict to set the footer of an amqp message.
+
 1.2.15 (2021-03-02)
 +++++++++++++++++++